+++ ZebVR/workers/image_filter.py
@@ -62,7 +62,7 @@ class ImageFilterWorker(WorkerNode):
             (data['index'], data['timestamp'], data['camera_timestamp'], image_processed),
             dtype=np.dtype([
                 ('index', int),
-                ('timestamp', np.float64),
+                ('timestamp', np.int64),
                 ('camera_timestamp', np.float64),
                 ('image', image_processed.dtype, image_processed.shape)
             ])
